Create Custom Linear Regression Function in R
This code is my final project for my Statistical Programming for Biostatisticians class from December 2024. The function takes outcome variables, predictors, and input data and generates parameter estimates, standard errors for each parameter estimate, and other helpful statistics. This function is consistent in performance to the lm() function in R.